neuronをモジュールに分割
